[fix] 이미지 업로드 시 cursor 위치 변경 / [refactor] editor 코드 스플리팅
